pins/classification/requirements.txt
appdirs==1.4.3
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.6
en-core-web-sm==1.2.0
murmurhash==0.26.4
nltk==3.2.2
numpy==1.12.1
packaging==16.8
pathlib==1.0.1
Pattern==2.6
plac==0.9.6
preshed==1.0.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
solrpy==0.9.7
spacy==1.7.3
termcolor==1.1.0
textrazor==1.3.0
thinc==6.5.2
toolz==0.8.2
tqdm==4.11.2
ujson==1.35
wrapt==1.10.10
